Refactor rate
Moved kuna_provider and stub provider to separate subfolder
Renamed service.go to rate.go
